+++ services/cloudfrontsigner/src/test/scala/facade/amazonaws/services/CloudfrontSignerTest.scala
@@ -1,11 +1,18 @@
-package net.exoego
+package facade.amazonaws.services
 
 import facade.amazonaws.services.cloudfront.signer
+import facade.amazonaws.services.cloudfront.signer.CloudFrontSigner
 import org.scalatest.funsuite.AnyFunSuite
 
 import scala.scalajs.js
 
 class CloudfrontSignerTest extends AnyFunSuite {
+  test("constructor") {
+    val o = new CloudFrontSigner("keyPairId", "privateKey")
+    assert(js.typeOf(o.asInstanceOf[js.Dynamic].getSignedCookie.asInstanceOf[js.Any]) === "function")
+    assert(js.typeOf(o.asInstanceOf[js.Dynamic].getSignedUrl.asInstanceOf[js.Any]) === "function")
+  }
+
   test("SignerOptionsWithPolicy.apply(PolicyStatement)") {
     val tPolicy = new signer.PolicyStatements(
       js.Array(
